fix(rl): address review (router multi-replica guard, init-block doc, validation/test cleanup)
